Rename GetImage - naming convention
