+++ packages/frontend/amp/lib/analytics.ts
@@ -23,7 +23,7 @@ export const analytics: (
     neilsenAPIID,
     domain,
 }) => [
-    `<amp-pixel src="//www.facebook.com/tr?id=${fbPixelaccount}&ev=PageView&noscript=1"></amp-pixel>`,
+    `<amp-pixel src="https://www.facebook.com/tr?id=${fbPixelaccount}&ev=PageView&noscript=1"></amp-pixel>`,
     `<amp-analytics config="https://ophan.theguardian.com/amp.json" data-credentials="include" ></amp-analytics>`,
     `<amp-analytics type="googleanalytics" id="google-analytics">
          <script type="application/json">
