+++ packages/frontend/amp/lib/image-fit.ts
